+++ cqp/attlist.h
@@ -27,7 +27,7 @@
 typedef struct _attrbuf {
   char *name;                           /**< name of the attribute */
   Attribute *attribute;                 /**< the relevant Attribute object, only valid if list_valid==1 */
-  int status;                           /**< This is user-settable; 0 on initialisation. */
+  int64_t status;                           /**< This is user-settable; 0 on initialisation. */
 
   struct _attrbuf * next;               /**< Next in chain. */
   struct _attrbuf * prev;               /**< Previous in chain */
@@ -36,8 +36,8 @@ typedef struct _attrbuf {
 /** The AttributeList object: holds a list of attributes */
 typedef struct _attlist {
 
-  int list_valid;                       /**< Are all the Attributes in this list valid? 0: check list */
-  int element_type;                     /**< One of the constants defined in attributes.h, format:ATT_x */
+  int64_t list_valid;                       /**< Are all the Attributes in this list valid? 0: check list */
+  int64_t element_type;                     /**< One of the constants defined in attributes.h, format:ATT_x */
 
   AttributeInfo *list;                  /**< Head of the linked list of attribute-info structures. */
 
@@ -45,30 +45,30 @@ typedef struct _attlist {
 
 /* ======================================== Allocation & Deallocation */
 
-AttributeList *NewAttributeList(int element_type);
+AttributeList *NewAttributeList(int64_t element_type);
 
-int DestroyAttributeList(AttributeList **list);
+int64_t DestroyAttributeList(AttributeList **list);
 
 /* ======================================== Adding and Removing */
 
 AttributeInfo *AddNameToAL(AttributeList *list,
                            char *name,
-                           int initial_status,
-                           int position);
+                           int64_t initial_status,
+                           int64_t position);
 
-int RemoveNameFromAL(AttributeList *list, char *name);
+int64_t RemoveNameFromAL(AttributeList *list, char *name);
 
-int NrOfElementsAL(AttributeList *list);
+int64_t NrOfElementsAL(AttributeList *list);
 
-int MemberAL(AttributeList *list, char *name);
+int64_t MemberAL(AttributeList *list, char *name);
 
 AttributeInfo *FindInAL(AttributeList *list, char *name);
 
-int RecomputeAL(AttributeList *list, Corpus *corpus, int initial_status);
+int64_t RecomputeAL(AttributeList *list, Corpus *corpus, int64_t initial_status);
 
-int VerifyList(AttributeList *list, 
+int64_t VerifyList(AttributeList *list, 
                Corpus *corpus,
-               int remove_illegal_entries);
+               int64_t remove_illegal_entries);
